requirements.txt
Flask==2.3.3
requests==2.31.0
websocket-client==1.6.4
pandas==2.1.3
gunicorn==21.2.0